PicoFaceD5: read the patches the way the MIDI implementation says to
Michael's suggestion, and the right one: stop reverse-engineering the
448-byte patch layout and check it against the D-50 MIDI implementation
manual, whose parameter list gives every offset, range and value table
outright. Five findings, three of them plain errors:

The LFO select is interleaved -- 0..5 means +1,-1,+2,-2,+3,-3, not the
+1,+2,+3,-1,-2,-3 this mapping guessed -- so every second PW, TVF and
TVA modulation rode the wrong LFO with the wrong sign. The WG LFO mode's
negative is value 2, not 3; and mode 3 is A&L, depth from aftertouch and
bender lever only, so those 65 partials now correctly get no fixed
vibrato instead of a full-depth one.

The pitch vibrato depth the partial block never contained is the common
block's P-Mod LFO Depth at offset 22, scaled to the service notes'
+-600 cents. Its absence was patched first with an invented quarter
swing (audibly detuned) and then with zero (audibly sterile); it is now
simply read.

The pitch envelope has no depth byte because its levels are the depth,
bipolar around 50 with a two-octave full scale -- the 2400-cent constant
an earlier revision zeroed in self-distrust was right all along, and is
restored.

And offset 2 is WG Pitch Keyfollow, seventeen ratios from -1 to 2 plus
the stretched tunings s1/s2, which this engine had silently ignored --
true for the 176 partials that sit on 1:1, wrong for the other 80, some
of which pin drum layers to fixed pitch or play half-scale by design.
Implemented pivoted on C4, with s1/s2 carried as 1.0 until someone
measures the stretch on hardware: a few cents of error where ignoring
the byte was whole scale degrees. TVF keyfollow (offset 15, same table)
comes along, so the filter tracks the keyboard too.

Verified on the factory bank: 57 of 64 sound the pressed key within 25
cents at C4; of the seven flagged, four sit inside their own patch's
intentional vibrato width (the single-bin measurement cannot tell
modulation from detune) and three are the inharmonic bells and mallets
whose strongest partial is off-grid by nature. Octave probes at C3/C5
land within 10 cents. Still deliberately unread: velocity ranges, TVA
bias, aftertouch and lever routes, bend mode -- listed so nobody
mistakes them for read.

+++ instruments/PicoFaceD5/include/d5_engine/d5_pcm_voice.h
@@ -38,7 +38,7 @@ struct PcmSampleRef {
 
 class PcmVoice {
 public:
-    void note_on(const PcmSampleRef& s, int note, float velocity,
+    void note_on(const PcmSampleRef& s, float note, float velocity,
                  const TvaEnvSpec& env, float sample_rate,
                  float detune = 1.0f) {
         s_ = s;
@@ -46,7 +46,7 @@ public:
         frac_ = 0.0f;
         active_ = s.data != nullptr && s.length > 0;
         gain_ = velocity;
-        const float f = 440.0f * std::pow(2.0f, (note - 69) / 12.0f);
+        const float f = 440.0f * std::pow(2.0f, (note - 69.0f) / 12.0f);
         // Unpitched material has nothing to transpose from: play it as stored.
         rate_ = (s.root_hz > 0.0f) ? f / s.root_hz : 1.0f;
         rate_ *= sample_rate > 0.0f ? (32000.0f / sample_rate) : 1.0f;
